CCL upgrade and native image config files (#75)
* Updated the submodule to point to the new commit

* Updated the submodule to point to the new commit

* chore: Build file for store

* Fix optional autowiring for WebServerApplicationContext

* Add tmp directory to .gitignore

This change prevents the temporary files in applications/store-build/tmp from being tracked by Git. It helps in keeping the repository clean and free from unnecessary files.

* Update Cardano node to version 10.1.2

Updated the download links and filenames for both AMD64 and ARM64 build scripts to retrieve version 10.1.2 of the Cardano node. This ensures the latest version is used in both architectures.

* Update Ogmios version in download script to v6.8.0

Upgraded the Ogmios version from v6.6.1 to v6.8.0 in the download-ogmios.sh script. This ensures that the latest features and fixes are included when downloading Ogmios.

* Add support for native store build and update configs

Introduce `yaci_store_mode=native` in the config and create a new Earthly target `store-native` for native store builds. Adjust the CLI Earthfile to pull the native store build based on the `STORE_NATIVE_BRANCH` when the build type is set to native.

* Add support for native mode execution of YaciStore

Introduce a new configuration property `yaci.store.mode` to enable running YaciStore in native mode. The service will now select either Java-based or native binary execution based on this property and the operating system. This change enhances the flexibility and performance optimization of the YaciStore service.

* Update version tag to 0.10.0-preview1

* Add git commit hash to build artifacts

Capture the current git commit hash in the `yaci-store` repository and save it as a build artifact. This provides a specific reference to the exact code version included in the build, improving traceability and debugging.

* Upgrade Cardano client libraries and refactor code.

Updated dependencies to newer versions of Cardano client and  additional interface mappings and serializers for native image configuration.

* Remove yaci-store application

The yaci-store application was redundant and not in use. This change cleans up unnecessary files from the repository.

* Remove .gitmodules file

* fix merge issue